third_party/blink/web_tests/fast/events/scale-and-scroll-body-expected.txt

PASS successfullyParsed is true

TEST COMPLETE

PASS window.document.scrollingElement.scrollTop is 100
PASS window.document.scrollingElement.scrollLeft is 100
PASS window.scrollX is 100
PASS window.scrollY is 100
